(2.6.12) Propose callback for exchange key
This callback was implemented because now you
would not receive exchange key until transaction
was completed. This callback allows consumer to
have exchange key early and store it somewhere
before tx completes.

It is important in scenario where tx execution
was interrupted and we want to retry it later (using
same txConfig from txExecutor callback). Previously
we could retry, but we would not have this key anymore